requirements.txt
Flask
python-dotenv
Werkzeug
PyMySQL
SQLAlchemy
Flask-SQLAlchemy
gunicorn